+++ Tests/MapboxDirectionsTests/MatchTests.swift
@@ -57,7 +57,7 @@ class MatchTests: XCTestCase {
         
         let tracePoints = match.tracepoints
         XCTAssertNotNil(tracePoints)
-        XCTAssertEqual(tracePoints.first!!.alternateCount, 0)
+        XCTAssertEqual(tracePoints.first!!.countOfAlternatives, 0)
         XCTAssertEqual(tracePoints.last!!.name, "West G Street")
 
         // confirming actual decoded values is important because the Directions API